Final correction in bonus part

+++ src/client_bonus.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   client.c                                           :+:      :+:    :+:   */
+/*   client_bonus.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nefimov <nefimov@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/11 14:55:12 by nefimov           #+#    #+#             */
-/*   Updated: 2025/02/18 14:30:53 by nefimov          ###   ########.fr       */
+/*   Updated: 2025/02/18 14:41:48 by nefimov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,12 @@ void	sigusr1_handler(int signum)
 	(void)signum;
 }
 
+void	sigusr2_handler(int signum)
+{
+	(void)signum;
+	ft_putstr_fd("String received\n", 1);
+}
+
 int	main(int argc, char *argv[])
 {
 	pid_t				s_pid;
@@ -26,9 +32,13 @@ int	main(int argc, char *argv[])
 	check_args(argc, argv, &s_pid);
 	if (signal(SIGUSR1, sigusr1_handler) == SIG_ERR)
 		return (1);
+	if (signal(SIGUSR2, sigusr2_handler) == SIG_ERR)
+		return (1);
 	len = ft_strlen(argv[2]);
 	kill(s_pid, SIGUSR1);
 	send_int(s_pid, len);
 	send_str(s_pid, argv[2], len);
+	pause();
+	usleep(PAUSE);
 	return (0);
 }
